pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-extractor"
version = "0.1.0"
description = "AI-powered document extraction system for college admissions"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Document Extractor Team"}
]

dependencies = [
    # Core Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    
    # Document Processing
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    
    # OCR
    "easyocr>=1.7.0",
    "pytesseract>=0.3.10",
    
    # AI/LLM
    "ollama>=0.1.0",
    "google-generativeai>=0.3.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    
    # Task Queue
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    
    # Data Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Utilities
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
doc-extractor = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
